refactor: Move all show methods in controllers to View
